pyproject.toml
[project]
name = "dicom-drl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pydicom>=3.0.1",
    "pylibjpeg>=2.1.0",
    "pylibjpeg-libjpeg>=2.3.0",
    "python-gdcm>=3.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
